+++ src/app/(user)/category/[idsessioncategory]/dataToFill.tsx
