+++ src/components/activityData/page/index.tsx
@@ -1,8 +1,8 @@
 'use client'
 
-import { Header } from "../../dashboard/header";
 import {  Container } from "@mui/system";
 import { ActivityDataForm } from "../Form";
+import { Header } from "../../dashboard/header";
 
 interface ActivityDataPageProps {
     handleConfirm: (type: string, value: string) => void;
